Fixed bug where video width and height wouldn't show properly

+++ BookBuilder/MainForm.cs
@@ -190,7 +190,7 @@ namespace BookBuilder
                 //Subract the page pictures location because the location of the video placeholder relative to the
                 //page picture is what matters.
                 double imageTop = (MainLayoutPanel.Size.Height - PagePicture.ImageRectangle.Size.Height) / 2.0;
-                double scale = (PagePicture.ImageRectangle.Width / currentPage.ImageWidth);
+                double scale = ((double)PagePicture.ImageRectangle.Width / currentPage.ImageWidth);
                 XPosBox.Text = ((int)((videoPlaceholder.Location.X - PagePicture.ImageRectangle.X - xVideoOffset) / scale)).ToString();
                 YPosBox.Text = ((int)((videoPlaceholder.Location.Y - imageTop - yVideoOffset) / scale)).ToString();
                 WidthBox.Text = ((int)(videoPlaceholder.Size.Width/ scale)).ToString();
@@ -296,8 +296,6 @@ namespace BookBuilder
                 PagePicture.Image = Image.FromFile(currentPage.SourcePageImageFileName);
                 currentPage.ImageHeight = PagePicture.Image.Height;
                 currentPage.ImageWidth = PagePicture.Image.Width;
-
-
             }
             else
             {
